+++ ecosystem/python/sdk/CHANGELOG.md
@@ -12,6 +12,10 @@ All notable changes to the Aptos Python SDK will be captured in this file. This 
 ## 0.6.4
 - Change sync client library from httpX to requests due to latency concerns.
 
+## 0.7.0
+- **[Breaking Change]**: Removed the `hex` function from `AccountAddress`. Instead of `addr.hex()` use `str(addr)`.
+- **[Breaking Change]**: The string representation of `AccountAddress` now conforms to [AIP-40](https://github.com/aptos-foundation/AIPs/blob/main/aips/aip-40.md).
+
 ## 0.6.2
 - Added custom header "x-aptos-client" to both sync/async RestClient